.depricated/requirements.txt
PyPDF2>=3.0.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
openai>=1.3.8
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.39.0
pillow>=10.2.0
pytesseract>=0.3.10
nltk>=3.8.1 